+++ src/main/kotlin/com/insyncwithfoo/pyrightls/configuration/application/Configurations.kt
@@ -6,4 +6,6 @@ import com.intellij.openapi.components.BaseState
 internal class Configurations : BaseState() {
     var alwaysUseGlobal by property(false)
     var globalExecutable by string(null)
+    var useEditorFont by property(false)
+    var addTooltipPrefix by property(false)
 }